Move the content codebase into the controller for templates.
